refactor: 쿠키에서 토큰을 추출하는 로직 분리
