+++ src/main/java/roomescape/config/WebConfig.java
@@ -5,8 +5,8 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import roomescape.argumentresolver.AuthArgumentResolver;
-import roomescape.interceptor.AuthInterceptor;
+import roomescape.infra.AuthArgumentResolver;
+import roomescape.infra.AuthInterceptor;
 
 @Configuration
 public class WebConfig implements WebMvcConfigurer {
