+++ src/main/java/roomescape/infra/AuthArgumentResolver.java
@@ -1,10 +1,6 @@
-package roomescape.argumentresolver;
+package roomescape.infra;
 
-import static roomescape.exception.ExceptionType.INVALID_TOKEN;
-
-import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
-import java.util.Arrays;
 import org.springframework.core.MethodParameter;
 import org.springframework.stereotype.Component;
 import org.springframework.web.bind.support.WebDataBinderFactory;
@@ -12,7 +8,6 @@ import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
 import roomescape.annotation.Auth;
-import roomescape.exception.RoomescapeException;
 import roomescape.service.TokenService;
 
 @Component
@@ -32,12 +27,7 @@ public class AuthArgumentResolver implements HandlerMethodArgumentResolver {
     public Object resolveArgument(MethodParameter parameter, ModelAndViewContainer mavContainer,
                                   NativeWebRequest webRequest, WebDataBinderFactory binderFactory) {
         HttpServletRequest nativeRequest = (HttpServletRequest) webRequest.getNativeRequest();
-        String token = Arrays.stream(nativeRequest.getCookies())
-                .filter(cookie -> cookie.getName().equals("token"))
-                .limit(1)
-                .findAny()
-                .map(Cookie::getValue)
-                .orElseThrow(() -> new RoomescapeException(INVALID_TOKEN));
+        String token = TokenExtractor.extractFrom(nativeRequest.getCookies());
         return tokenService.findUserIdFromToken(token);
     }
 }
