+++ src/main/java/roomescape/infra/AuthInterceptor.java
@@ -1,12 +1,9 @@
-package roomescape.interceptor;
+package roomescape.infra;
 
-import static roomescape.exception.ExceptionType.INVALID_TOKEN;
 import static roomescape.exception.ExceptionType.PERMISSION_DENIED;
 
-import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import java.util.Arrays;
 import org.springframework.stereotype.Component;
 import org.springframework.web.servlet.HandlerInterceptor;
 import roomescape.domain.Role;
@@ -27,12 +24,7 @@ public class AuthInterceptor implements HandlerInterceptor {
 
     @Override
     public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) {
-        String token = Arrays.stream(request.getCookies())
-                .filter(cookie -> cookie.getName().equals("token"))
-                .limit(1)
-                .findAny()
-                .map(Cookie::getValue)
-                .orElseThrow(() -> new RoomescapeException(INVALID_TOKEN));
+        String token = TokenExtractor.extractFrom(request.getCookies());
         long userIdFromToken = tokenService.findUserIdFromToken(token);
         UserInfo userInfo = memberService.findByUserId(userIdFromToken);
         if (!userInfo.role().equals(Role.ADMIN.name())) {
